강의 종료되면 평가 모달 뜨게 하기

+++ data/src/main/java/org/softwaremaestro/data/review_create/model/ReviewReqDto.kt
@@ -3,6 +3,6 @@ package org.softwaremaestro.data.review_create.model
 import com.google.gson.annotations.SerializedName
 
 data class ReviewReqDto(
-    @SerializedName("rating") val rating: Float,
+    @SerializedName("rating") val rating: Int,
     @SerializedName("comment") val comment: String
 )
