+++ domain/src/main/java/org/softwaremaestro/domain/review_create/entity/ReviewVO.kt
@@ -2,6 +2,6 @@ package org.softwaremaestro.domain.review_create.entity
 
 data class ReviewVO(
     val tutoringId: String,
-    val rating: Float,
+    val rating: Int,
     val comment: String
 )
